Adds zsh configuration.
Since different shells are being used on different systems, an ignore file was added which ignores
the configuration file of the shell that is not being in use, or all of them if no shell is specified.

Currently, `.chezmoi.shell` to "zsh" will install .zshrc while setting it to "fish" will install config.fish.