Corrects the uses of templates variables.
Incorrect syntax for templates variables was used before, in particular, custom
variables under the toml configuration were accessed with `.chezmoi.var` instead
of `.var`.

The uses were corrected all around.
Furthermore, it was noted that chezmoi provides some variables automatically.
For this reason, every use of the `os` custom variable was changed to
`chezmoi.osRelease.name`, deprecating the need for the custom variable in the
first place.

`run_once_install_starship.sh` was modified to require the shell to be fish to
perform the installation, as powerline10K is used on zsh instead.